SwiftRTCuda: do not vend CUDA as SwiftRTCuda
CUDA is a separate library, not vended by SwiftRTCuda, although it does
re-export it for users.  Split the CUDA into a local module definition
(which is SPM friendly) and build atop of that.  This allows the removal
of the superfluous `srt_cuda_api.h` header, which is now inlined into
`SwiftRTCuda.h` as an umbrella header.

The only issue that remains to be solved here is the modulemap
redistribution - the modulemap contains build specific configuration
(namely the location of the CUDA headers).  As the modulemap is not
installed next to the headers, it is not possible to use a path relative
location and modulemaps are currently not relocatable.